requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.1.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
facenet-pytorch==2.6.0
faiss-cpu==1.11.0
filelock==3.17.0
filetype==1.2.0
fonttools==4.55.8
frozenlist==1.6.0
fsspec==2025.2.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.7
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
langchain-text-splitters==0.3.8
langsmith==0.3.42
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.8.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.11.0.86
openpyxl==3.1.5
orjson==3.10.18
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==10.2.0
pillow_heif==0.22.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
propcache==0.3.1
psutil==6.1.1
pure_eval==0.2.3
py-cpuinfo==9.0.0
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.1
pytesseract==0.3.10
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywin32==310
PyYAML==6.0.2
pyzmq==26.3.0
rapidfuzz==3.5.2
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.54
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
stack-data==0.6.3
sympy==1.13.1
tenacity==9.1.2
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
torch==2.2.2
torchaudio==2.6.0
torchvision==0.17.2
tornado==6.4.2
touch==2020.12.3
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.133
ultralytics-thop==2.0.14
urllib3==2.3.0
wcwidth==0.2.13
yarl==1.20.0
zstandard==0.23.0
